tests/wpt/tests/dom/nodes/getElementsByClassName-02.htm

<!doctype html>
<html class="a
b">
 <head>
  <title>document.getElementsByClassName(): also simple</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body class="a
">
  <div id="log"></div>
  <script> test(function() {assert_array_equals(document.getElementsByClassName("a\n"), [document.documentElement, document.body])}) </script>
 </body>
</html>
